Move disaggregation init override to common mixin
